Rename assert sub-package and removed duplicate methods from 0x.js's assert module
